Refactor the implementation of the controller's probe and migration
Before this PR, the API and controller is very hard to add test cases
since we don't design any interface for this purpose. And many
components are designed as an async mode by default which is not
necessary.

So we have resolved those issues by using the sync mode and add more
interface to make it easier to mock and test.