Introduece etcd's raft to implement the KV store
Currently, the controller must depend on the external service like
zookeeper/etcd to store its data. And it's not so friendly for users
since they need to setup another cluster before running the controller
server. To make this process simple, we would like to introduce the raft
engine to save and replicate the data.